moves wrapper for usdb file parsing to mw.py
